Added support.py
support.py is used to define 2 import functions (folder import and csv layout of map). support.py will be needed again if we want to switch to marwans map to generate folder and csv of the map. Ive created an issue for map integration but it has no status atp. we'll see what we do with it.