Magic functionality ( 3:44:03 - 4:00:19 )
This sets up the commands and execution of magic. 

The player.py takes the inputs for allowing magic and to change spell. 

The settings.py contains the dictionary of the spells. 

The ui.py has the magic spell currently equipped next to the weapons icon. 

The level.py just setups the function for magic , as of this stage it only prints the data of magic in dictionary.