Attack Mechanics ( 4:00:19 - 4:57:00 )
This allows the player to deal Damage to the enemy and also receive damage. 

The level.py allows grass to be instantly destroyed when attacked and manipulates health of player and enemy. 

The weapon.py just updates the weapons to a sprite type of 'weapons' 

The enemy.py makes the enemy flinch back and flicker when attacked , also allowing collision with 'weapons' . Makes sure the enemy is dead 

The entity.py just updates to include a function to allow the flicker of enemy and player .

The player.py is updated to allow flickering , flinching  and Makes sure the damage dealt is inclusive of weapon damage and player stats.