Particle effects ( 5:32:50 - 5:56:25)
This adds a more logical and seamless effect to the game interactions.

The enemy.py allows the death of an enemy to have an animation. 

The level.py updates to show the animation of leaves when they are hit. And the effect when a player gets damage from enemy. 

The particles.py has the dictionary of all the animations of the particles. It contains the grass particle creation along with making the leaves mirrored. Player and enemy particle effects also created in this file.