Spells Execution ( 5:56:25 - 6:26:35 )
This Push just allows the spells to work . The heal , heals the player . The flame deals damage to the enemy. 

The magic.py sets the requirements ( sufficient energy ) for both the spell and adds the heal effect to the health. This also creates the particles for both the spells. The flame particles are set here to follow the direction that the player is facing. 

The particles.py just updates to include the sprite type 'magic'. 

The player.py includes the energy recovery function , which replenishes energy constantly at a set rate. It also defines how much Damage the flame deals.

The enemy.py allows for the flame particles to deal damage to the enemies by the amount defined in player.py. 

The level.py updates to call the 'heal' and 'flame' functions from magic.py