Upgrade Mechanic ( 6:26:35 - 7:24:26 )
This Allows the player to gain experience by killing enemies and then use the experience to upgrade player stats. 

The enemy.py just gives the player experience to be gained after enemy death. 

The settings.py includes the upgrade menu's Colour data's. 

The level.py adds the enemy exp to player exp. It also Imports from a new file ( upgrade.py ) . And it has a function to pause the game and then when paused to call the upgrade menu display. 

The main.py has another if statement which checks if key pressed ( u ) and then calls the level.py function to pause the game. 

The player.py has two new functions which return the cost to upgrade each stat and by how much to upgrade. 

The upgrade.py creates / displays the upgrade menu and takes keyboard inputs to upgrade stats .